+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Changed column and Parry default tilt standard deviation from 1.0
+  to 0.5, and same for Parry rotation standard deviation
+
 ## 4.1.0 - 2022-07-23
 
 ### Changed